+++ tests/testthat/test-abnormal_by_worst_grade_by_baseline.R
@@ -11,20 +11,26 @@ test_that("h_adsl_adlb_merge_using_worst_flag generates missing for patients wit
 
   result <- h_adsl_adlb_merge_using_worst_flag(adsl_f, adlb_f, worst_flag = c("WGRHIFL" = "Y"))
 
-  result_matrix <- matrix(c(as.matrix(result)), 6L, 6L)
+  result_matrix <- matrix(
+    c(as.matrix(result[c("USUBJID", "ARMCD", "AVISIT", "PARAMCD", "PARAM", "ATOXGR", "BTOXGR")])),
+    6L,
+    7L)
 
   expected_matrix <- structure(
     c(rep(c("AB12345-CHN-3-id-128", "AB12345-CHN-15-id-262"), each = 3L),
       rep(c("ARM A", "ARM C"), each = 3L),
       c("WEEK 3 DAY 22", "WEEK 5 DAY 36", "WEEK 3 DAY 22", "<Missing>", "<Missing>", "<Missing>"),
       rep(c("ALT", "CRP", "IGA"), 2L),
+      rep(
+        c("Alanine Aminotransferase Measurement", "C-Reactive Protein Measurement", "Immunoglobulin A Measurement"),
+        2L),
       c("2", "0", "0", "<Missing>", "<Missing>", "<Missing>"),
       c("0", "0", "3", "<Missing>", "<Missing>", "<Missing>")
     ),
-    .Dim = c(6L, 6L)
+    .Dim = c(6L, 7L)
   )
 
-  expected_matrix <- matrix(c(expected_matrix), 6L, 6L)
+  expected_matrix <- matrix(c(expected_matrix), 6L, 7L)
 
   expect_identical(result_matrix, expected_matrix)
 })
@@ -37,21 +43,27 @@ test_that("h_adsl_adlb_merge_using_worst_flag generates missing for patients mis
 
   result <- h_adsl_adlb_merge_using_worst_flag(adsl_f, adlb_f, worst_flag = c("WGRHIFL" = "Y"))
 
-  result_matrix <- matrix(c(as.matrix(result)), 6L, 6L)
+  result_matrix <- matrix(
+    c(as.matrix(result[c("USUBJID", "ARMCD", "AVISIT", "PARAMCD", "PARAM", "ATOXGR", "BTOXGR")])),
+    6L,
+    7L)
 
   expected_matrix <- structure(
     c(
       rep(c("AB12345-CHN-15-id-262", "AB12345-CHN-3-id-128"), each = 3L),
       rep(c("ARM C", "ARM A"), each = 3L),
       c("WEEK 5 DAY 36", "WEEK 1 DAY 8", "WEEK 4 DAY 29", "WEEK 3 DAY 22", "WEEK 5 DAY 36", "WEEK 3 DAY 22"),
-      rep(c("ALT", "CRP", "IGA"), 2),
+      rep(c("ALT", "CRP", "IGA"), 2L),
+      rep(
+        c("Alanine Aminotransferase Measurement", "C-Reactive Protein Measurement", "Immunoglobulin A Measurement"),
+        2L),
       c("0", "0", "0", "2", "0", "0"),
       c("<Missing>", "<Missing>", "<Missing>", "0", "0", "3")
     ),
-    .Dim = c(6L, 6L)
+    .Dim = c(6L, 7L)
   )
 
-  expected_matrix <- matrix(c(expected_matrix), 6L, 6L)
+  expected_matrix <- matrix(c(expected_matrix), 6L, 7L)
 
   expect_identical(result_matrix, expected_matrix)
 })
@@ -64,7 +76,10 @@ test_that("h_adsl_adlb_merge_using_worst_flag generates missing for patients mis
 
   result <- h_adsl_adlb_merge_using_worst_flag(adsl_f, adlb_f, worst_flag = c("WGRHIFL" = "Y"))
 
-  result_matrix <- matrix(c(as.matrix(result)), 6L, 6L)
+  result_matrix <- matrix(
+    c(as.matrix(result[c("USUBJID", "ARMCD", "AVISIT", "PARAMCD", "PARAM", "ATOXGR", "BTOXGR")])),
+    6L,
+    7L)
 
   expected_matrix <- structure(
     c(
@@ -72,18 +87,24 @@ test_that("h_adsl_adlb_merge_using_worst_flag generates missing for patients mis
       rep(c("ARM C", "ARM A"), each = 3L),
       c("WEEK 5 DAY 36", "WEEK 1 DAY 8", "WEEK 4 DAY 29", "WEEK 3 DAY 22", "WEEK 5 DAY 36", "WEEK 3 DAY 22"),
       rep(c("ALT", "CRP", "IGA"), 2L),
+      rep(
+        c("Alanine Aminotransferase Measurement", "C-Reactive Protein Measurement", "Immunoglobulin A Measurement"),
+        2L),
       c("<Missing>", "<Missing>", "<Missing>", "2", "0", "0"),
       c("0", "2", "0", "0", "0", "3")
     ),
-    .Dim = c(6L, 6L)
+    .Dim = c(6L, 7L)
   )
 
-  expected_matrix <- matrix(c(expected_matrix), 6L, 6L)
+  expected_matrix <- matrix(c(expected_matrix), 6L, 7L)
 
   expect_identical(result_matrix, expected_matrix)
 })
 
-test_that("h_adsl_adlb_merge_using_worst_flag generates missing for patients without any post-baseline values flagged as the worst", { #nolint
+test_that(
+  "h_adsl_adlb_merge_using_worst_flag generates missing for patients without
+  any post-baseline values flagged as the worst",
+  { #nolint
   adsl_f <- adsl %>% filter(USUBJID %in% c("AB12345-CHN-3-id-128", "AB12345-CHN-15-id-262"))
   adlb_f <- adlb %>% filter(USUBJID %in% c("AB12345-CHN-3-id-128", "AB12345-CHN-15-id-262"))
 
@@ -91,7 +112,10 @@ test_that("h_adsl_adlb_merge_using_worst_flag generates missing for patients wit
 
   result <- h_adsl_adlb_merge_using_worst_flag(adsl_f, adlb_f, worst_flag = c("WGRHIFL" = "Y"))
 
-  result_matrix <- matrix(c(as.matrix(result)), 6L, 6L)
+  result_matrix <- matrix(
+    c(as.matrix(result[c("USUBJID", "ARMCD", "AVISIT", "PARAMCD", "PARAM", "ATOXGR", "BTOXGR")])),
+    6L,
+    7L)
 
   expected_matrix <- structure(
     c(
@@ -99,13 +123,16 @@ test_that("h_adsl_adlb_merge_using_worst_flag generates missing for patients wit
       rep(c("ARM A", "ARM C"), each = 3L),
       c("WEEK 3 DAY 22", "WEEK 5 DAY 36", "WEEK 3 DAY 22", "<Missing>", "<Missing>", "<Missing>"),
       rep(c("ALT", "CRP", "IGA"), 2L),
+      rep(
+        c("Alanine Aminotransferase Measurement", "C-Reactive Protein Measurement", "Immunoglobulin A Measurement"),
+        2L),
       c("2", "0", "0", "<Missing>", "<Missing>", "<Missing>"),
       c("0", "0", "3", "<Missing>", "<Missing>", "<Missing>")
     ),
-    .Dim = c(6L, 6L)
+    .Dim = c(6L, 7L)
   )
 
-  expected_matrix <- matrix(c(expected_matrix), 6L, 6L)
+  expected_matrix <- matrix(c(expected_matrix), 6L, 7L)
 
   expect_identical(result_matrix, expected_matrix)
 })
@@ -323,7 +350,10 @@ test_that("h_adsl_adlb_merge_using_worst_flag generates missing and by visit lab
 
   result <- h_adsl_adlb_merge_using_worst_flag(adsl_f, adlb_f, worst_flag = c("WGRHIVFL" = "Y"), by_visit = TRUE)
 
-  result_matrix <- matrix(c(as.matrix(result)), 30L, 6L)
+  result_matrix <- matrix(
+    c(as.matrix(result[c("USUBJID", "ARMCD", "AVISIT", "PARAMCD", "PARAM", "ATOXGR", "BTOXGR")])),
+    30L,
+    7L)
 
   expected_matrix <- structure(
     c(
@@ -352,7 +382,21 @@ test_that("h_adsl_adlb_merge_using_worst_flag generates missing and by visit lab
       "ALT", "ALT", "ALT", "CRP", "CRP", "CRP", "CRP", "CRP", "IGA",
       "IGA", "IGA", "IGA", "IGA", "ALT", "ALT", "ALT", "ALT", "ALT",
       "CRP", "CRP", "CRP", "CRP", "CRP", "IGA", "IGA", "IGA", "IGA",
-      "IGA", "-2", "-3", "2", "0", "-3", "-1", "0", "0", "0", "0",
+      "IGA", "Alanine Aminotransferase Measurement", "Alanine Aminotransferase Measurement",
+      "Alanine Aminotransferase Measurement", "Alanine Aminotransferase Measurement",
+      "Alanine Aminotransferase Measurement", "C-Reactive Protein Measurement",
+      "C-Reactive Protein Measurement", "C-Reactive Protein Measurement",
+      "C-Reactive Protein Measurement", "C-Reactive Protein Measurement",
+      "Immunoglobulin A Measurement", "Immunoglobulin A Measurement", "Immunoglobulin A Measurement",
+      "Immunoglobulin A Measurement", "Immunoglobulin A Measurement",
+      "Alanine Aminotransferase Measurement", "Alanine Aminotransferase Measurement",
+      "Alanine Aminotransferase Measurement", "Alanine Aminotransferase Measurement",
+      "Alanine Aminotransferase Measurement", "C-Reactive Protein Measurement",
+      "C-Reactive Protein Measurement", "C-Reactive Protein Measurement",
+      "C-Reactive Protein Measurement", "C-Reactive Protein Measurement",
+      "Immunoglobulin A Measurement", "Immunoglobulin A Measurement", "Immunoglobulin A Measurement",
+      "Immunoglobulin A Measurement", "Immunoglobulin A Measurement",
+      "-2", "-3", "2", "0", "-3", "-1", "0", "0", "0", "0",
       "0", "0", "0", "0", "0", "<Missing>", "<Missing>", "<Missing>",
       "<Missing>", "<Missing>", "<Missing>", "<Missing>", "<Missing>",
       "<Missing>", "<Missing>", "<Missing>", "<Missing>", "<Missing>",
@@ -362,10 +406,10 @@ test_that("h_adsl_adlb_merge_using_worst_flag generates missing and by visit lab
       "<Missing>", "<Missing>", "<Missing>", "<Missing>", "<Missing>",
       "<Missing>", "<Missing>", "<Missing>"
     ),
-    .Dim = c(30L, 6L)
+    .Dim = c(30L, 7L)
   )
 
-  expected_matrix <- matrix(c(expected_matrix), 30L, 6L)
+  expected_matrix <- matrix(c(expected_matrix), 30L, 7L)
 
   expect_identical(result_matrix, expected_matrix)
 })